diskwriter: fix test scope
